requirements.txt
# requirements.txt
# This file contains the dependencies for the HealthPredict AI application.
# Ensure to keep this file updated with the latest versions of the libraries useed in the project.  
# Use `pip install -r requirements.txt` to install all dependencies.
# Core
pandas>=2.0
numpy>=1.24
scikit-learn>=1.4
matplotlib>=3.7
altair>=5.0
joblib>=1.3
pyyaml>=6.0

# App
streamlit>=1.35
shap>=0.45

# NLP
transformers>=4.41
torch>=2.2
sentencepiece>=0.2

# HTTP
requests>=2.31
